+++ src/components/composite/UsersList/UsersList.tsx
@@ -3,6 +3,7 @@ import "./UsersList.css";
 import User from "../../primitives/User/User";
 
 type UserType = {
+  twitch_id: number | string
   display_name: string;
   profile_image_url: string;
 };
@@ -12,7 +13,7 @@ function UsersList({ users }: {users: UserType[] | undefined}) {
     <div className="userList">
       {users && users.map((user, i) => {
         return (
-          <div className="user">
+          <div className="user" key={user.twitch_id}>
             <User
               userName={user.display_name}
               profileImgURL={user.profile_image_url}
