+++ src/main.tsx
@@ -5,7 +5,6 @@ import { RouterProvider, createBrowserRouter } from "react-router-dom";
 import "./index.css";
 
 import ViewersPage from "./pages/ViewersPage/ViewersPage.tsx";
-import ConnectMainPage from "./pages/ConnectMainPage/ConnectMainPage.tsx";
 import HomePage from "./pages/HomePage/HomePage.tsx";
 import DashboardPage from "./pages/DashboardPage/DashboardPage.tsx";
 import Visualizer from "./test/Visualizer.tsx";
@@ -23,10 +22,10 @@ const router = createBrowserRouter([
     path: "/dashboard",
     element: <DashboardPage />,
   },
-  /*{
+  {
     path: "/viewers",
     element: <ViewersPage />,
-  },*/
+  },
 ]);
 
 createRoot(document.getElementById("root")!).render(
