+++ src/pages/ViewersPage/ViewersPage.css
@@ -1,15 +1,19 @@
-.base {
-	gap: 5px;
-	display: block;
+@media (min-width: 800px) {
+	.mainSection {
+		display: flex;
+		justify-content: center;
+		align-items: center;
+	}
+
+	.modDiv {
+		min-width: 500px;
+	}
 }
 
-.H1 {
-	font-size: 1.1rem;
-}
-
-.mainSection {
-	margin: 1.8rem 1.8rem;
-	display: flex;
-	flex-direction: column;
-	gap: 1.3rem;
+.modDiv {
+	max-width: 700px;
+	margin: 30px 20px;
+	border-radius: 15px;
+	border: #353538 solid 2px;
+	overflow: hidden;
 }
